requirements.txt
streamlit>=1.35.0
python-pptx>=0.6.23
requests>=2.31.0
